refactor: QueryDsl 커스텀 인터페이스 생성하여 기존 repository에서 상속하는 것으로 리팩토링
